+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/permission/MiniAppCustomPermissionCache.kt
@@ -1,5 +1,6 @@
 package com.rakuten.tech.mobile.miniapp.permission
 
+import android.annotation.SuppressLint
 import android.content.Context
 import android.content.SharedPreferences
 import androidx.annotation.VisibleForTesting
@@ -17,34 +18,72 @@ internal class MiniAppCustomPermissionCache(context: Context) {
         "com.rakuten.tech.mobile.miniapp.custom.permissions.cache", Context.MODE_PRIVATE
     )
 
+    @VisibleForTesting(otherwise = VisibleForTesting.NONE)
+    fun doesDataExist(miniAppId: String) = prefs.contains(miniAppId)
+
     /**
      * Reads the grant results from SharedPreferences.
      * @param [miniAppId] the key provided to find the stored results per MiniApp
      * @return [MiniAppCustomPermission] an object to contain the results per MiniApp
      * if data has been stored in cache, otherwise default value.
      */
+    @SuppressLint("RestrictedApi")
+    @SuppressWarnings("NestedBlockDepth")
     fun readPermissions(miniAppId: String): MiniAppCustomPermission {
-        if (prefs.contains(miniAppId)) {
-            return try {
-                Gson().fromJson(
+        val defaultValue = defaultDeniedList(miniAppId)
+        return if (doesDataExist(miniAppId)) {
+            try {
+                val cachedPermission: MiniAppCustomPermission = Gson().fromJson(
                     prefs.getString(miniAppId, ""),
                     object : TypeToken<MiniAppCustomPermission>() {}.type
                 )
+                val cachedPairs = cachedPermission.pairValues.toMutableList()
+
+                // detect any new change with comparing cached permissions and defaultDeniedList
+                // change means added new permission / removed existing permission from defaultDeniedList
+                val defaultPairs = defaultValue.pairValues
+                val changedPermissions = (defaultPairs + cachedPairs).groupBy { it.first.type }
+                    .filter { it.value.size == 1 }
+                    .flatMap { it.value }
+
+                return if (changedPermissions.isNotEmpty()) {
+                    if (cachedPairs.size < defaultPairs.size) {
+                        val filteredValue =
+                            MiniAppCustomPermission(miniAppId, cachedPairs + changedPermissions)
+                        applyStoringPermissions(filteredValue)
+                        filteredValue
+                    } else {
+                        cachedPairs.removeAll { (first) ->
+                            first.type in changedPermissions.groupBy { it.first.type }
+                        }
+                        val filteredValue = MiniAppCustomPermission(miniAppId, cachedPairs)
+                        applyStoringPermissions(filteredValue)
+                        filteredValue
+                    }
+                } else {
+                    val filteredValue = MiniAppCustomPermission(miniAppId, cachedPairs)
+                    applyStoringPermissions(filteredValue)
+                    filteredValue
+                }
             } catch (e: Exception) {
-                defaultDeniedList(miniAppId)
+                // if there is any exception, just return the default value
+                defaultValue
             }
+        } else {
+            // if value hasn't been found in SharedPreferences, save the value
+            applyStoringPermissions(defaultValue)
+            defaultValue
         }
-        return defaultDeniedList(miniAppId)
     }
 
     /**
      * Stores the grant results to SharedPreferences.
      * @param [miniAppCustomPermission] an object to contain the results per MiniApp.
      */
+    @SuppressLint("RestrictedApi")
     fun storePermissions(
         miniAppCustomPermission: MiniAppCustomPermission
     ) {
-        val cached = readPermissions(miniAppCustomPermission.miniAppId).pairValues
         val supplied = miniAppCustomPermission.pairValues.toMutableList()
 
         // Remove any unknown permission parameter from HostApp.
@@ -52,28 +91,59 @@ internal class MiniAppCustomPermissionCache(context: Context) {
             first.type == MiniAppCustomPermissionType.UNKNOWN.type
         }
 
+        val miniAppId = miniAppCustomPermission.miniAppId
+        val allPermissions = prepareAllPermissionsToStore(miniAppId, supplied)
+        applyStoringPermissions(MiniAppCustomPermission(miniAppId, allPermissions))
+    }
+
+    @SuppressWarnings("PrintStackTrace")
+    @VisibleForTesting(otherwise = VisibleForTesting.NONE)
+    fun applyStoringPermissions(miniAppCustomPermission: MiniAppCustomPermission) {
         try {
-            val jsonToStore: String = Gson().toJson(
-                MiniAppCustomPermission(
-                    miniAppCustomPermission.miniAppId,
-                    prepareAllPermissionsToStore(cached, supplied)
-                )
-            )
+            val jsonToStore: String = Gson().toJson(orderByDefaultList(miniAppCustomPermission))
             prefs.edit().putString(miniAppCustomPermission.miniAppId, jsonToStore).apply()
         } catch (e: Exception) {
-            e.message.toString()
+            e.printStackTrace()
         }
     }
 
-    /**
-     * Prepares all custom permissions per MiniApp by comparing the cached and supplied
-     * custom permissions with replacing old grant results with new grant results.
-     */
-    @VisibleForTesting
-    internal fun prepareAllPermissionsToStore(
-        cached: List<Pair<MiniAppCustomPermissionType, MiniAppCustomPermissionResult>>,
+    @VisibleForTesting(otherwise = VisibleForTesting.NONE)
+    fun orderByDefaultList(miniAppCustomPermission: MiniAppCustomPermission): MiniAppCustomPermission {
+        val miniAppId = miniAppCustomPermission.miniAppId
+        val defaultTypesOrder = mutableListOf<MiniAppCustomPermissionType>()
+        defaultDeniedList(miniAppId).pairValues.forEach {
+            defaultTypesOrder.add(it.first)
+        }
+
+        val currentTypesOrder = mutableListOf<MiniAppCustomPermissionType>()
+        miniAppCustomPermission.pairValues.forEach {
+            currentTypesOrder.add(it.first)
+        }
+
+        val expectedTypesOrder = currentTypesOrder.map {
+            defaultTypesOrder.indexOf(it)
+        }.sorted().map { value -> defaultTypesOrder[value] }
+
+        val orderedPair =
+            mutableListOf<Pair<MiniAppCustomPermissionType, MiniAppCustomPermissionResult>>()
+
+        expectedTypesOrder.forEachIndexed { index, type ->
+            miniAppCustomPermission.pairValues.find {
+                it.first == type
+            }?.let { orderedPair.add(index, it) }
+        }
+
+        return MiniAppCustomPermission(miniAppId, orderedPair)
+    }
+
+    @VisibleForTesting(otherwise = VisibleForTesting.NONE)
+    fun prepareAllPermissionsToStore(
+        miniAppId: String,
         supplied: List<Pair<MiniAppCustomPermissionType, MiniAppCustomPermissionResult>>
     ): List<Pair<MiniAppCustomPermissionType, MiniAppCustomPermissionResult>> {
+        // retrieve permissions by comparing cached and supplied (from HostApp) permissions
+        // readPermissions already filters the changed permissions from defaultDeniedList
+        val cached = readPermissions(miniAppId).pairValues
         val combined = (cached + supplied).toMutableList()
         combined.removeAll { (first) ->
             first.type in supplied.groupBy { it.first.type }
@@ -81,8 +151,12 @@ internal class MiniAppCustomPermissionCache(context: Context) {
         return combined + supplied
     }
 
-    @VisibleForTesting
-    internal fun defaultDeniedList(miniAppId: String): MiniAppCustomPermission {
+    /**
+     * Note: Update this default list when adding or removing a custom permission,
+     * [MiniAppCustomPermissionCache] should automatically handle the value.
+     */
+    @VisibleForTesting(otherwise = VisibleForTesting.NONE)
+    fun defaultDeniedList(miniAppId: String): MiniAppCustomPermission {
         return MiniAppCustomPermission(
             miniAppId,
             listOf(
@@ -96,6 +170,10 @@ internal class MiniAppCustomPermissionCache(context: Context) {
                 ),
                 Pair(
                     MiniAppCustomPermissionType.CONTACT_LIST,
+                    MiniAppCustomPermissionResult.DENIED
+                ),
+                Pair(
+                    MiniAppCustomPermissionType.LOCATION,
                     MiniAppCustomPermissionResult.DENIED
                 )
             )