+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/permission/PermissionsUtils.kt
@@ -8,6 +8,7 @@ fun toReadableName(type: MiniAppCustomPermissionType?): String {
         MiniAppCustomPermissionType.USER_NAME -> "User Name"
         MiniAppCustomPermissionType.CONTACT_LIST -> "Contact List"
         MiniAppCustomPermissionType.PROFILE_PHOTO -> "Profile Photo"
+        MiniAppCustomPermissionType.LOCATION -> "Device Location"
         else -> "Unknown"
     }
 }